build debian instead.

+++ platform/broadcom/sswsyncd/dsserve.cpp
@@ -147,7 +147,7 @@
          else
          {
              /* print orphaned message to the stdout */
-             printf("%.*s", rc, data);
+             printf("%.*s", (int)rc, data);
              fflush(stdout);
          }
      }
